+++ metric_test.go
@@ -20,11 +20,11 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"go.opentelemetry.io/otel/metric"
-	"go.opentelemetry.io/otel/metric/embedded"
+	"go.opentelemetry.io/otel/metric/metricembed"
 	"go.opentelemetry.io/otel/metric/noop"
 )
 
-type testMeterProvider struct{ embedded.MeterProvider }
+type testMeterProvider struct{ metricembed.MeterProvider }
 
 var _ metric.MeterProvider = &testMeterProvider{}