+++ sdk/metric/instrument.go
@@ -20,8 +20,8 @@ import (
 	"fmt"
 
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/metric/embedded"
 	"go.opentelemetry.io/otel/metric/instrument"
+	"go.opentelemetry.io/otel/metric/metricembed"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/internal"
@@ -173,12 +173,12 @@ type streamID struct {
 type instrumentImpl[N int64 | float64] struct {
 	aggregators []internal.Aggregator[N]
 
-	embedded.Float64Counter
-	embedded.Float64UpDownCounter
-	embedded.Float64Histogram
-	embedded.Int64Counter
-	embedded.Int64UpDownCounter
-	embedded.Int64Histogram
+	metricembed.Float64Counter
+	metricembed.Float64UpDownCounter
+	metricembed.Float64Histogram
+	metricembed.Int64Counter
+	metricembed.Int64UpDownCounter
+	metricembed.Int64Histogram
 }
 
 var _ instrument.Float64Counter = (*instrumentImpl[float64])(nil)
@@ -222,9 +222,9 @@ type float64Observable struct {
 	instrument.Float64Observable
 	*observable[float64]
 
-	embedded.Float64ObservableCounter
-	embedded.Float64ObservableUpDownCounter
-	embedded.Float64ObservableGauge
+	metricembed.Float64ObservableCounter
+	metricembed.Float64ObservableUpDownCounter
+	metricembed.Float64ObservableGauge
 }
 
 var _ instrument.Float64ObservableCounter = float64Observable{}
@@ -241,9 +241,9 @@ type int64Observable struct {
 	instrument.Int64Observable
 	*observable[int64]
 
-	embedded.Int64ObservableCounter
-	embedded.Int64ObservableUpDownCounter
-	embedded.Int64ObservableGauge
+	metricembed.Int64ObservableCounter
+	metricembed.Int64ObservableUpDownCounter
+	metricembed.Int64ObservableGauge
 }
 
 var _ instrument.Int64ObservableCounter = int64Observable{}