+++ sdk/metric/meter.go
@@ -22,8 +22,8 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/internal/global"
 	"go.opentelemetry.io/otel/metric"
-	"go.opentelemetry.io/otel/metric/embedded"
 	"go.opentelemetry.io/otel/metric/instrument"
+	"go.opentelemetry.io/otel/metric/metricembed"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/metric/internal"
 )
@@ -33,7 +33,7 @@ import (
 // produced by an instrumentation scope will use metric instruments from a
 // single meter.
 type meter struct {
-	embedded.Meter
+	metricembed.Meter
 
 	scope instrumentation.Scope
 	pipes pipelines
@@ -267,7 +267,7 @@ func (m *meter) RegisterCallback(f metric.Callback, insts ...instrument.Observab
 }
 
 type observer struct {
-	embedded.Observer
+	metricembed.Observer
 
 	float64 map[observablID[float64]]struct{}
 	int64   map[observablID[int64]]struct{}
@@ -361,7 +361,7 @@ func (r observer) ObserveInt64(o instrument.Int64Observable, v int64, a ...attri
 	oImpl.observe(v, a)
 }
 
-type noopRegister struct{ embedded.Registration }
+type noopRegister struct{ metricembed.Registration }
 
 func (noopRegister) Unregister() error {
 	return nil
@@ -419,7 +419,7 @@ func (p int64ObservProvider) callback(i int64Observable, f instrument.Int64Callb
 }
 
 type int64Observer struct {
-	embedded.Int64Observer
+	metricembed.Int64Observer
 	int64Observable
 }
 
@@ -451,7 +451,7 @@ func (p float64ObservProvider) callback(i float64Observable, f instrument.Float6
 }
 
 type float64Observer struct {
-	embedded.Float64Observer
+	metricembed.Float64Observer
 	float64Observable
 }